# Removing file.close() statements