Resolves #3. This PR removes the `file.close()` statements from `DataSet` methods. This is important because, as-is, `DataSet` closes the file it is reading _every time_ it slices data. This could be costly when doing a lot of data slicing as the numerous data i/o actions slows a script down. Instead, the file remains open as an attribute of `DataSet` unless `DataSet.close()` is called. A nice "unintended" consequence of this change is that the file name is [no longer needed when slicing data](https://github.com/stompsjo/RadClass/blob/9ddb3d72b1fab664e9190481c023efc45a519270/RadClass/DataSet.py#L85).

A new test (`test_close()`) has been included to ensure the method is functional, although I'm not sure how necessary it is?

There is some [interesting](https://stackoverflow.com/questions/865115/how-do-i-correctly-clean-up-a-python-object) [discussion](https://stackoverflow.com/questions/8582076/class-wrapper-around-file-proper-way-to-close-file-handle-when-no-longer-refe) about using a `.close()` method instead of a `__enter__` and `__exit__` combination. As mentioned in this Stack Overflow pages, I think a `.close()` method is appropriate here since `DataSet.file` should never be called outside of `DataSet` and therefore will be closed no matter what when `DataSet` is closed. Instead, the purpose of `.close()` is to provide parent classes a way of "recycling" a child `DataSet` object to analyze several data files and should be completely optional. This seems like an appropriate workflow for this project.