@stompsjo 2021-04-05
I did a quick test, calculating the hourly gross count-rate for a 1-month file and comparing the time to completion before and after the change:

Before: 96.28 s
After: 95.68 s

So a 0.6% speed-up. Not too much, but this script is only slicing the data 2,678,400 s / 3600 s = 744 times. You can imagine a more significant speed up for an analysis that slices, say, 44,640 times (if it was calculating the 1-second gross count rate).